Split scenes / components
