+++ src/App.js
@@ -2,7 +2,7 @@ import 'babel-polyfill';
 import 'semantic-ui-css/semantic.min.css';
 import React from 'react';
 
-import MainMenu from './Game/MainMenu';
+import MainMenu from './scenes/Game/MainMenu';
 
 const App = () => <MainMenu />;
