+++ src/components/Board/components/Square/components/Piece/black-advisor-50px.svg.png
