+++ src/components/Board/components/Square/components/Piece/black-cannon-50px.svg.png
