+++ src/components/Board/components/Square/components/Piece/black-chariot-50px.svg.png
