+++ src/components/Board/components/Square/components/Piece/black-elephant-50px.svg.png
