+++ src/components/Board/components/Square/components/Piece/black-horse-50px.svg.png
