+++ src/components/Board/components/Square/components/Piece/black-soldier-50px.svg.png
