+++ src/components/Board/components/Square/components/Piece/images.js
