+++ src/components/Board/components/Square/components/Piece/index.js
