+++ src/components/Board/components/Square/components/Piece/red-advisor-50px.svg.png
