+++ src/components/Board/components/Square/components/Piece/red-cannon-50px.svg.png
