+++ src/components/Board/components/Square/components/Piece/red-chariot-50px.svg.png
