+++ src/components/Board/components/Square/components/Piece/red-elephant-50px.svg.png
