+++ src/components/Board/components/Square/components/Piece/red-horse-50px.svg.png
