+++ src/components/Board/components/Square/components/Piece/red-soldier-50px.svg.png
