+++ src/scenes/Game/GameListItem.js
