+++ src/scenes/Game/reducers.js
@@ -1,7 +1,7 @@
 import update from 'immutability-helper';
 import { useReducer } from 'react';
-import XiangqiBoard, * as logic from '../logic';
-import { Color } from '../logic/constants';
+import XiangqiBoard, * as logic from '../../logic';
+import { Color } from '../../logic/constants';
 
 // Initial State
